replace logging with handler
